requirements.txt
chronos-forecasting==1.5.0
numpy==1.26.4
ollama==0.4.7
pandas==2.1.4
matplotlib==3.7.5
seaborn==0.13.2
scikit-learn==1.4.2
tsfresh==0.21.0
imbalanced-learn==0.13.0
torch==2.5.0
optuna==4.2.1
info-nce-pytorch==0.1.4
ollama==0.4.7
jupyter
ipykernel==6.29.5
notebook==7.3.3
fastparquet==2024.11.0